+++ internal/parser/parser.go
@@ -8942,6 +8942,15 @@ type StrictOptions struct {
 	// legacy reasons such that "document.all != null" is false even though it's
 	// not "null" or "undefined".
 	NullishCoalescing bool
+
+	// Loose:  "class Foo { foo = 1 }" => "class Foo { constructor() { this.foo = 1; } }"
+	// Strict: "class Foo { foo = 1 }" => "class Foo { constructor() { __publicField(this, 'foo', 1); } }"
+	//
+	// The disadvantage of strictness here is code bloat and performance. The
+	// advantage is following the class field specification accurately. For
+	// example, loose mode will incorrectly trigger setter methods while strict
+	// mode won't.
+	ClassFields bool
 }
 
 type ParseOptions struct {